requirements.txt
Django==4.2
ipdb==0.13.13
flake8==6.0.0
django-environ==0.10.0
pillow==9.4.0
faker==18.7.0
django-widget-tweaks==1.4.12
weasyprint==58.1
django-lifecycle==1.0.0
celery[redis]==5.2.7
django-celery-results==2.5.0
django-extensions==3.2.1
django-celery-beat==2.5.0
django-rosetta==0.9.9
django-silk==5.0.3
django-debug-toolbar==4.0.0
pytest==7.3.1
pytest-django==4.5.2
pytest-cov==4.0.0
pytest-factoryboy==2.5.1
freezegun==1.2.2
pytest-mock==3.10.0
gunicorn==20.1.0
beautifulsoup4==4.12.2
python-slugify==8.0.1
django-filter==23.2
djangorestframework==3.14.0
